Introduce `ErrorPresentableView` protocol for providing custom error views.
