+++ Source/Turbo/Navigator/Helpers/ErrorHandling/ErrorPresentableView.swift
@@ -1,0 +1,7 @@
+import Foundation
+import SwiftUI
+
+public protocol ErrorPresentableView: View {
+    var error: Error { get }
+    var handler: ErrorPresenter.Handler? { get }
+}
